// app-serveuse/src/types/database.types.ts
export interface Table {
  id: string;
  numero: number;
  capacite: number;
  statut: 'libre' | 'occupee' | 'commande_en_attente';
  etablissement_id: string;
}